🚚 Move: statistics 폴더 재구성
