--- src/components/MonthlyStatisticsContainer/MonthlyStatisticsContainer.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-interface MonthlyStatisticsContainerProps {}
-
-const MonthlyStatisticsContainer: React.FC<MonthlyStatisticsContainerProps> = ({}) => {
-  return (
-    <div>
-      <p>월간 통계다~~~~</p>
-    </div>
-  );
-};
-
-export default MonthlyStatisticsContainer;
