+++ src/components/statistics/monthly/MonthlyStatistics.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+interface MonthlyStatisticsProps {}
+
+const MonthlyStatistics: React.FC<MonthlyStatisticsProps> = ({}) => {
+  return (
+    <div>
+      <p>월간 통계다~~~~</p>
+    </div>
+  );
+};
+
+export default MonthlyStatistics;
