+++ src/components/statistics/weekly/WeeklyCompletion/CompletionBarGraph/CompletionBarGraph.stories.tsx
