+++ src/components/statistics/weekly/WeeklyCompletion/CompletionBarGraph/CompletionBarGraph.tsx
