+++ src/components/statistics/weekly/WeeklyCompletion/CompletionText/CompletionText.tsx
