+++ src/components/statistics/weekly/WeeklyRanking/WeeklyOthers/WeeklyOthers.tsx
