+++ src/components/statistics/weekly/WeeklyStatAction/WeeklyStatAction.stories.tsx
