+++ src/components/statistics/weekly/WeeklyStatAction/WeeklyStatAction.tsx
